+++ test/neuralnetworks/core_test.clj
@@ -2,7 +2,15 @@
   (:require [clojure.test :refer :all]
             [neuralnetworks.core :as nn]
             [clojure.core.matrix :as m]
-            [neuralnetworks.sigmoid-fn :as sigmoid-fn]))
+            [neuralnetworks.optimizer.stopping-conditions :refer :all]))
+
+(deftest test-randomize-theta-dimensions
+  (let [thetas-no-hidden-layer (nn/randomize-thetas 2 [] 1)
+        thetas-one-hidden-layer (nn/randomize-thetas 2 [3] 1)
+        thetas-two-hidden-layers (nn/randomize-thetas 2 [1 1] 1)]
+    (is (= [[1 3]] (map m/shape thetas-no-hidden-layer)))
+    (is (= [[3 3] [1 4]] (map m/shape thetas-one-hidden-layer)))
+    (is (= [[1 3] [1 2] [1 2]] (map m/shape thetas-two-hidden-layers)))))
 
 (deftest test-train-and
   (let [input (m/array [[0 0]
@@ -16,14 +24,14 @@
         theta-hidden-layer-output (m/array [[0.453169 -1.037344 -0.837474 0.985438]])
         thetas [theta-input-hidden-layer theta-hidden-layer-output]
         output (m/array [[0] [0] [0] [1]])
-        instance (nn/new-instance input thetas output {})]
-    (nn/train! instance)
+        instance (nn/new-instance input thetas output :classification {})]
+    (nn/train! instance [(max-error 0.01)])
     (is (m/equals output (nn/predict instance input) 0.1))))
 
 (deftest test-train-xor
-  (let [input (m/array [[-1 -1]
-                        [-1 1]
-                        [1 -1]
+  (let [input (m/array [[0 0]
+                        [0 1]
+                        [1 0]
                         [1 1]])
         ; thetas values are generated from `(nn/randomize-thetas 2 [3] 1)`
         theta-input-hidden-layer (m/array [[-0.453452 0.510620 -0.773563]
@@ -31,18 +39,7 @@
                                            [-0.534591 0.786677 0.756434]])
         theta-hidden-layer-output (m/array [[0.453169 -1.037344 -0.837474 0.985438]])
         thetas [theta-input-hidden-layer theta-hidden-layer-output]
-        output (m/array [[-1] [1] [1] [-1]])
-        instance (nn/new-instance input thetas output {:sigmoid-fn sigmoid-fn/hyperbolic-tangent
-                                                       :optimizer (neuralnetworks.optimizer.gradient-descent/gradient-descent 1 1 [(neuralnetworks.optimizer.stopping-conditions/max-iterations 1)])})]
-        ;instance (nn/new-instance input thetas output {})]
-    (time (nn/train! instance))
-    (prn (nn/predict instance input))))
-    ;(is (m/equals output (nn/predict instance input) 0.1))))
-
-(deftest test-randomize-theta-dimensions
-  (let [thetas-no-hidden-layer (nn/randomize-thetas 2 [] 1)
-        thetas-one-hidden-layer (nn/randomize-thetas 2 [3] 1)
-        thetas-two-hidden-layers (nn/randomize-thetas 2 [1 1] 1)]
-    (is (= [[1 3]] (map m/shape thetas-no-hidden-layer)))
-    (is (= [[3 3] [1 4]] (map m/shape thetas-one-hidden-layer)))
-    (is (= [[1 3] [1 2] [1 2]] (map m/shape thetas-two-hidden-layers)))))
+        output (m/array [[0] [1] [1] [0]])
+        instance (nn/new-instance input thetas output :classification {})]
+    (nn/train! instance [(max-error 0.01)])
+    (is (m/equals output (nn/predict instance input) 0.1))))